wrong address fix fix fix

+++ commands/texmodcmd.cpp
@@ -93,7 +93,7 @@ void TexModCmd::undo()
     for (u32 i = 0; i < m_mod_count; i++)
     {
         tmp = *((u8*)((intptr_t)m_mod[i].first + img));
-        *((u8*)((int)m_mod[i].first + img)) = m_mod[i].second;
+        *((u8*)((intptr_t)m_mod[i].first + img)) = m_mod[i].second;
         m_mod[i].second = tmp;
     }
